requirements.txt
anyio==4.9.0
asgiref==3.6.0
certifi==2025.4.26
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==22.1
pytz==2023.3
rapidfuzz==3.9.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.3
typing_extensions==4.14.0
tzdata==2023.3
